Refactor DB entities to use int64 and time_t types & new persistence DTOs have been added
Updated all DAOs, entities, and SQL queries to use int64_t for IDs and time_t for date fields, replacing string representations. Added player_number to Play table and logic, introduced DTOs for joined queries, and implemented new functions for retrieving joined data. This improves type safety, consistency, and enables more efficient date handling throughout the backend.
Some persistence DTOs have been added, which is mainly used to recover the player nickname, as well as a function that recovers all plays by round.

+++ source-code/backend/src/entities/round_entity.h
@@ -22,8 +22,8 @@ typedef enum {
 } RoundStatus;
 
 typedef struct {
-    int id_round;
-    int id_game;
+    int64_t id_round;
+    int64_t id_game;
     RoundStatus state;
     int64_t duration;
     char board[BOARD_MAX];
